requirements.txt
# Core dependencies
anthropic>=0.40.0
gspread>=6.1.0
oauth2client>=4.1.3
requests>=2.32.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
thefuzz[speedup]>=0.22.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Optional for Teams notifications
pymsteams>=0.2.2
